ajustes

+++ db/migrate/20191118232224_create_mov_estoques.rb
@@ -2,7 +2,7 @@ class CreateMovEstoques < ActiveRecord::Migration[5.2]
   def change
     create_table :mov_estoques do |t|
       t.date :data
-      t.integer :quantidade
+      t.float :quantidade
       t.belongs_to :pessoa, foreign_key: true
       t.belongs_to :operacao, foreign_key: true
       t.belongs_to :produto, foreign_key: true
